refactor: Refactor restaurants.js
將restaurants.js依MVC架構重構
